+++ Backend 1 - Intro a Java/Guia 12 - Herencia/EjerciciosExtra/G12_EjExtra3_Hoteles/src/entities/Hotel5.java
@@ -44,8 +44,8 @@ public final class Hotel5 extends Hotel4 {
     }
 
     @Override
-    public float precioHabitacion() {
-        float precio = super.precioHabitacion();
+    public double precioHabitacion() {
+        double precio = super.precioHabitacion();
         precio += (15 * cantLimosinas);
         
         return precio;
